# apis/v1alpha1/openapi.yaml
openapi: 3.0.0
info:
  title: Portal API
  description: PortalAPIのOpenAPI Spec
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers: []
security:
  - BearerAuth: []
  - CookieAuth: []
tags:
- name: "applications"
  description: "アプリケーション関連のAPI"
- name: "health"
  description: "ヘルスチェック関連のAPI"
- name: "auth"
  description: "認証関連のAPI"
externalDocs: {}
paths:
  /auth/github/login:
    get:
      tags:
        - "auth"
      summary: "GitHub OAuth Login"
      description: "GitHub OAuthログインを開始するAPI"
      operationId: "GitHubOAuthLogin"
      security: []  # 認証不要
      responses:
        '302':
          description: "GitHub OAuth認証サイトにリダイレクト"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/github/callback:
    get:
      tags:
        - "auth"
      summary: "GitHub OAuth Callback"
      description: "GitHub OAuthコールバック処理API"
      operationId: "GitHubOAuthCallback"
      security: []  # 認証不要
      parameters:
        - name: "code"
          in: "query"
          description: "Authorization code"
          required: true
          schema:
            type: "string"
        - name: "state"
          in: "query"
          description: "OAuth state parameter"
          required: true
          schema:
            type: "string"
      responses:
        '200':
          description: "認証成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        '400':
          description: "認証エラー"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/csrf-token:
    get:
      tags:
        - "auth"
      summary: "Get CSRF Token"
      description: "CSRFトークンを取得するAPI。セッションCookie認証が必要。"
      operationId: "GetCSRFToken"
      security:
        - CookieAuth: []
      responses:
        '200':
          description: "CSRFトークン取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CSRFTokenResponse"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      tags:
        - "auth"
      summary: "Logout"
      description: "ログアウト処理API"
      operationId: "Logout"
      security:
        - CookieAuth: []
      responses:
        '200':
          description: "ログアウト成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/refresh:
    post:
      tags:
        - "auth"
      summary: "Token Refresh"
      description: "JWT リフレッシュ処理API"
      operationId: "RefreshToken"
      security: []  # リフレッシュトークンで認証
      requestBody:
        description: "リフレッシュトークン"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required:
                - refresh_token
      responses:
        '200':
          description: "トークンリフレッシュ成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        '401':
          description: "無効なリフレッシュトークン"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /health/liveness:
    get:
      tags:
        - "health"
      summary: "Get Liveness Status"
      description: "liveness statusを取得するAPI"
      operationId: "GetHealthLiveness"
      security: []  # 認証不要
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "Liveness statusの取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthCheckStatus"
  /health/readiness:
    get:
      tags:
        - "health"
      summary: "Get Readiness Status"
      description: "readiness statusを取得するAPI"
      operationId: "GetHealthReadiness"
      security: []  # 認証不要
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "Readiness statusの取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthCheckStatus"
  /v1alpha1/applications:
    get:
      tags:
        - "applications"
      summary: "Get Applications"
      description: "アプリケーション一覧を取得するAPI"
      operationId: "GetApplications"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "アプリケーション一覧の取得成功"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      tags:
        - "applications"
      summary: "Create Application"
      description: "アプリケーションを作成するAPI"
      operationId: "CreateApplication"
      requestBody:
        description: "作成するアプリケーションの情報"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '201':
          description: "アプリケーションの作成成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /v1alpha1/applications/{name}:
    get:
      tags:
        - "applications"
      summary: "Get Application"
      description: "特定のアプリケーションを取得するAPI"
      operationId: "GetApplication"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "アプリケーションの取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    put:
      tags:
        - "applications"
      summary: "Update Application"
      description: "特定のアプリケーションを更新するAPI"
      operationId: "UpdateApplication"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      requestBody:
        description: "更新するアプリケーションの情報"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
      responses:
        '200':
          description: "アプリケーションの更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        '404':
          description: "アプリケーションが見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - "applications"
      summary: "Delete Application"
      description: "特定のアプリケーションを削除するAPI"
      operationId: "DeleteApplication"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      responses:
        '204':
          description: "アプリケーションの削除成功"
        '404':
          description: "アプリケーションが見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1alpha1/applications/{name}/secret:
    get:
      tags:
        - "applications"
      summary: "Get Application Secret"
      description: "特定のアプリケーションのシークレットを取得するAPI"
      operationId: "GetApplicationSecret"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "アプリケーションシークレットの取得成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
    post:
      tags:
        - "applications"
      summary: "Create Application Secret"
      description: "特定のアプリケーションのシークレットを作成するAPI"
      operationId: "CreateApplicationSecret"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      requestBody:
        description: "作成するシークレットの情報"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '201':
          description: "アプリケーションシークレットの作成成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
    put:
      tags:
        - "applications"
      summary: "Update Application Secret"
      description: "特定のアプリケーションのシークレットを更新するAPI"
      operationId: "UpdateApplicationSecret"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      requestBody:
        description: "更新するシークレットの情報"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSecretRequest"
      responses:
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '200':
          description: "アプリケーションシークレットの更新成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
    delete:
      tags:
        - "applications"
      summary: "Delete Application Secret"
      description: "特定のアプリケーションのシークレットを削除するAPI"
      operationId: "DeleteApplicationSecret"
      parameters:
        - name: "name"
          in: "path"
          description: "アプリケーション名"
          required: true
          schema:
            type: "string"
      responses:
        '204':
          description: "アプリケーションシークレットの削除成功"
        '404':
          description: "シークレットが見つからない"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: "デフォルトのレスポンス"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "GitHub PAT or Installation Token"
    CookieAuth:
      type: apiKey
      in: cookie
      name: portal_session
      description: "OAuth JWT Cookie"
  schemas:
    CSRFTokenResponse:
      type: object
      properties:
        csrf_token:
          type: string
      required:
        - csrf_token
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
      required:
        - code
        - message
    HealthCheckStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Application:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        display_id:
          type: string
        repository_url:
          type: string
        appconfig_path:
          type: string
        appconfig_branch:
          type: string
      required:
        - id
        - name
        - display_id
        - repository_url
        - appconfig_path
        - appconfig_branch
    CreateApplicationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        repository_url:
          type: string
        appconfig_path:
          type: string
        appconfig_branch:
          type: string
      required:
        - name
        - repository_url
        - appconfig_path
        - appconfig_branch
    SecretItem:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value
    Secret:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SecretItem"
      required:
        - id
        - items
    UpdateApplicationRequest:
      type: object
      properties:
        repository_url:
          type: string
        appconfig_path:
          type: string
        appconfig_branch:
          type: string
      required:
        - repository_url
        - appconfig_path
        - appconfig_branch
    CreateSecretRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SecretItem"
      required:
        - items
    User:
      type: object
      properties:
        id:
          type: string
        login:
          type: string
        name:
          type: string
        email:
          type: string
        teams:
          type: array
          items:
            type: string
      required:
        - id
        - login
        - teams
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
        expires_at:
          type: string
          format: date-time
      required:
        - success
        - user
        - expires_at
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
      required:
        - access_token
        - refresh_token
        - expires_at