# pkg/apis/v1alpha1/api/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"

				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "csrf-token"

					if l := len("csrf-token"); len(elem) >= l && elem[0:l] == "csrf-token" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetCSRFTokenRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				case 'g': // Prefix: "github/"

					if l := len("github/"); len(elem) >= l && elem[0:l] == "github/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "callback"

						if l := len("callback"); len(elem) >= l && elem[0:l] == "callback" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGitHubOAuthCallbackRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					case 'l': // Prefix: "login"

						if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleGitHubOAuthLoginRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}

					}

				case 'l': // Prefix: "logout"

					if l := len("logout"); len(elem) >= l && elem[0:l] == "logout" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleLogoutRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

				case 'r': // Prefix: "refresh"

					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleRefreshTokenRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}

				}

			case 'h': // Prefix: "health/"

				if l := len("health/"); len(elem) >= l && elem[0:l] == "health/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "liveness"

					if l := len("liveness"); len(elem) >= l && elem[0:l] == "liveness" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetHealthLivenessRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				case 'r': // Prefix: "readiness"

					if l := len("readiness"); len(elem) >= l && elem[0:l] == "readiness" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleGetHealthReadinessRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}

				}

			case 'v': // Prefix: "v1alpha1/applications"

				if l := len("v1alpha1/applications"); len(elem) >= l && elem[0:l] == "v1alpha1/applications" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "GET":
						s.handleGetApplicationsRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleCreateApplicationRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "name"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch r.Method {
						case "DELETE":
							s.handleDeleteApplicationRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleGetApplicationRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "PUT":
							s.handleUpdateApplicationRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET,PUT")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/secret"

						if l := len("/secret"); len(elem) >= l && elem[0:l] == "/secret" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleDeleteApplicationSecretRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "GET":
								s.handleGetApplicationSecretRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "POST":
								s.handleCreateApplicationSecretRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							case "PUT":
								s.handleUpdateApplicationSecretRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "DELETE,GET,POST,PUT")
							}

							return
						}

					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"

			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"

				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'c': // Prefix: "csrf-token"

					if l := len("csrf-token"); len(elem) >= l && elem[0:l] == "csrf-token" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetCSRFTokenOperation
							r.summary = "Get CSRF Token"
							r.operationID = "GetCSRFToken"
							r.operationGroup = ""
							r.pathPattern = "/auth/csrf-token"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'g': // Prefix: "github/"

					if l := len("github/"); len(elem) >= l && elem[0:l] == "github/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "callback"

						if l := len("callback"); len(elem) >= l && elem[0:l] == "callback" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GitHubOAuthCallbackOperation
								r.summary = "GitHub OAuth Callback"
								r.operationID = "GitHubOAuthCallback"
								r.operationGroup = ""
								r.pathPattern = "/auth/github/callback"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					case 'l': // Prefix: "login"

						if l := len("login"); len(elem) >= l && elem[0:l] == "login" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "GET":
								r.name = GitHubOAuthLoginOperation
								r.summary = "GitHub OAuth Login"
								r.operationID = "GitHubOAuthLogin"
								r.operationGroup = ""
								r.pathPattern = "/auth/github/login"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}

					}

				case 'l': // Prefix: "logout"

					if l := len("logout"); len(elem) >= l && elem[0:l] == "logout" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = LogoutOperation
							r.summary = "Logout"
							r.operationID = "Logout"
							r.operationGroup = ""
							r.pathPattern = "/auth/logout"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "refresh"

					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "POST":
							r.name = RefreshTokenOperation
							r.summary = "Token Refresh"
							r.operationID = "RefreshToken"
							r.operationGroup = ""
							r.pathPattern = "/auth/refresh"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'h': // Prefix: "health/"

				if l := len("health/"); len(elem) >= l && elem[0:l] == "health/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "liveness"

					if l := len("liveness"); len(elem) >= l && elem[0:l] == "liveness" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetHealthLivenessOperation
							r.summary = "Get Liveness Status"
							r.operationID = "GetHealthLiveness"
							r.operationGroup = ""
							r.pathPattern = "/health/liveness"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				case 'r': // Prefix: "readiness"

					if l := len("readiness"); len(elem) >= l && elem[0:l] == "readiness" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch method {
						case "GET":
							r.name = GetHealthReadinessOperation
							r.summary = "Get Readiness Status"
							r.operationID = "GetHealthReadiness"
							r.operationGroup = ""
							r.pathPattern = "/health/readiness"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}

				}

			case 'v': // Prefix: "v1alpha1/applications"

				if l := len("v1alpha1/applications"); len(elem) >= l && elem[0:l] == "v1alpha1/applications" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						r.name = GetApplicationsOperation
						r.summary = "Get Applications"
						r.operationID = "GetApplications"
						r.operationGroup = ""
						r.pathPattern = "/v1alpha1/applications"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = CreateApplicationOperation
						r.summary = "Create Application"
						r.operationID = "CreateApplication"
						r.operationGroup = ""
						r.pathPattern = "/v1alpha1/applications"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"

					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "name"
					// Match until "/"
					idx := strings.IndexByte(elem, '/')
					if idx < 0 {
						idx = len(elem)
					}
					args[0] = elem[:idx]
					elem = elem[idx:]

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							r.name = DeleteApplicationOperation
							r.summary = "Delete Application"
							r.operationID = "DeleteApplication"
							r.operationGroup = ""
							r.pathPattern = "/v1alpha1/applications/{name}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							r.name = GetApplicationOperation
							r.summary = "Get Application"
							r.operationID = "GetApplication"
							r.operationGroup = ""
							r.pathPattern = "/v1alpha1/applications/{name}"
							r.args = args
							r.count = 1
							return r, true
						case "PUT":
							r.name = UpdateApplicationOperation
							r.summary = "Update Application"
							r.operationID = "UpdateApplication"
							r.operationGroup = ""
							r.pathPattern = "/v1alpha1/applications/{name}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/secret"

						if l := len("/secret"); len(elem) >= l && elem[0:l] == "/secret" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch method {
							case "DELETE":
								r.name = DeleteApplicationSecretOperation
								r.summary = "Delete Application Secret"
								r.operationID = "DeleteApplicationSecret"
								r.operationGroup = ""
								r.pathPattern = "/v1alpha1/applications/{name}/secret"
								r.args = args
								r.count = 1
								return r, true
							case "GET":
								r.name = GetApplicationSecretOperation
								r.summary = "Get Application Secret"
								r.operationID = "GetApplicationSecret"
								r.operationGroup = ""
								r.pathPattern = "/v1alpha1/applications/{name}/secret"
								r.args = args
								r.count = 1
								return r, true
							case "POST":
								r.name = CreateApplicationSecretOperation
								r.summary = "Create Application Secret"
								r.operationID = "CreateApplicationSecret"
								r.operationGroup = ""
								r.pathPattern = "/v1alpha1/applications/{name}/secret"
								r.args = args
								r.count = 1
								return r, true
							case "PUT":
								r.name = UpdateApplicationSecretOperation
								r.summary = "Update Application Secret"
								r.operationID = "UpdateApplicationSecret"
								r.operationGroup = ""
								r.pathPattern = "/v1alpha1/applications/{name}/secret"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}

					}

				}

			}

		}
	}
	return r, false
}